requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
cfgv==3.4.0
click==8.1.8
distlib==0.3.9
environs==14.1.1
fastapi==0.115.12
filelock==3.18.0
flake8==7.2.0
h11==0.14.0
identify==2.6.9
idna==3.10
isort==6.0.1
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pre_commit==4.2.0
pycodestyle==2.13.0
pydantic==2.11.3
pydantic_core==2.33.1
pyflakes==3.3.2
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.0
virtualenv==20.30.0
fastapi
uvicorn
gspread
google-auth
httpx~=0.28.1